+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - On Windows, fix bug where `RedrawRequested` would only get emitted every other iteration of the event loop.
 - On X11, fix deadlock on window state when handling certain window events.
 - `WindowBuilder` now implements `Default`.
+- On macOS, fix `CursorMoved` event reporting the cursor position using logical coordinates.
 
 # 0.20.0 (2020-01-05)